requirements.txt
# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Data scraping and processing
beautifulsoup4>=4.12.0
praw>=7.7.0
feedparser>=6.0.0
newspaper3k>=0.2.8
requests>=2.31.0
lxml>=4.9.0

# Sentiment Analysis
textblob>=0.17.0
vaderSentiment>=3.3.0
nltk>=3.8.0

# Data manipulation and visualization - using flexible versions for Python 3.13
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
